--- a/src/main/java/org/opensearch/searchrelevance/dao/JudgmentCacheDao.java
+++ b/src/main/java/org/opensearch/searchrelevance/dao/JudgmentCacheDao.java
@@ -150,7 +150,10 @@ public class JudgmentCacheDao {
                 SearchHit hit = response.getHits().getHits()[0];
             }
             listener.onResponse(response);
-        }, e -> { listener.onFailure(e); });
+        }, e -> {
+            LOGGER.debug("Cache lookup failed for docId: {} - continuing without cache", documentId);
+            listener.onFailure(e);
+        });
 
         return searchRelevanceIndicesManager.listDocsBySearchRequest(searchSourceBuilder, JUDGMENT_CACHE, wrappedListener);
     }