--- a/src/main/java/org/opensearch/searchrelevance/judgments/JudgmentsProcessorFactory.java
+++ b/src/main/java/org/opensearch/searchrelevance/judgments/JudgmentsProcessorFactory.java
@@ -13,6 +13,7 @@ import org.opensearch.searchrelevance.dao.QuerySetDao;
 import org.opensearch.searchrelevance.dao.SearchConfigurationDao;
 import org.opensearch.searchrelevance.ml.MLAccessor;
 import org.opensearch.searchrelevance.model.JudgmentType;
+import org.opensearch.threadpool.ThreadPool;
 import org.opensearch.transport.client.Client;
 
 public class JudgmentsProcessorFactory {
@@ -20,8 +21,8 @@ public class JudgmentsProcessorFactory {
     private final QuerySetDao querySetDao;
     private final SearchConfigurationDao searchConfigurationDao;
     private final JudgmentCacheDao judgmentCacheDao;
-
     private final Client client;
+    private final ThreadPool threadPool;
 
     @Inject
     public JudgmentsProcessorFactory(
@@ -29,18 +30,27 @@ public class JudgmentsProcessorFactory {
         QuerySetDao querySetDao,
         SearchConfigurationDao searchConfigurationDao,
         JudgmentCacheDao judgmentCacheDao,
-        Client client
+        Client client,
+        ThreadPool threadPool
     ) {
         this.mlAccessor = mlAccessor;
         this.querySetDao = querySetDao;
         this.searchConfigurationDao = searchConfigurationDao;
         this.judgmentCacheDao = judgmentCacheDao;
         this.client = client;
+        this.threadPool = threadPool;
     }
 
     public BaseJudgmentsProcessor getProcessor(JudgmentType type) {
         return switch (type) {
-            case LLM_JUDGMENT -> new LlmJudgmentsProcessor(mlAccessor, querySetDao, searchConfigurationDao, judgmentCacheDao, client);
+            case LLM_JUDGMENT -> new LlmJudgmentsProcessor(
+                mlAccessor,
+                querySetDao,
+                searchConfigurationDao,
+                judgmentCacheDao,
+                client,
+                threadPool
+            );
             case UBI_JUDGMENT -> new UbiJudgmentsProcessor(client);
             case IMPORT_JUDGMENT -> new ImportJudgmentsProcessor(client);
             default -> throw new IllegalArgumentException("Unsupported judgment type: " + type);