--- a/src/main/java/org/opensearch/searchrelevance/ml/ChunkResult.java
+++ b/src/main/java/org/opensearch/searchrelevance/ml/ChunkResult.java
@@ -49,6 +49,10 @@ public class ChunkResult {
         this.failedChunks = new HashMap<>(failedChunks);
     }
 
+    public int getChunkIndex() {
+        return chunkIndex;
+    }
+
     public int getSuccessfulChunksCount() {
         return succeededChunks.size();
     }