### Description
According to RFC 159, implementing task scheduler methods can significantly improve the efficiency of LLM judgment processing. 

### Potential Issues:
LLM Judgments, current issues starting from class LlmJudgmentsProcessor:
- Sequential processing of multiple queries in generateLLMJudgments()
- Blocking search operations across multiple search configurations
- Sequential LLM prediction calls with actionGet() blocking patterns
- Cache lookups performed sequentially per document

### Key Transformations:

> Sequential processing of multiple queries :white_check_mark:


- LlmJudgmentTaskManager with concurrent processing
  - queries processed in parallel with semaphore-based rate limiting

> Blocking search operations :white_check_mark:


- Search operations are now asynchronous within each task
 - No more sequential blocking across search configurations

> Sequential LLM prediction calls :white_check_mark:


- LLM calls are now concurrent per query via task manager
- Each query processes independently in parallel

> Sequential cache lookups :white_check_mark:


- Cache operations are asynchronous with graceful fallback
- Cache failures don’t block processing

:white_check_mark: Current Architecture is Non-Blocking:
```
LlmJudgmentsProcessor
├── Upfront cache index creation (async, non-blocking)
├── LlmJudgmentTaskManager.scheduleTasksAsync()
│   ├── {querySetSize} concurrent tasks (semaphore-limited to ~24)
│   ├── Each task: processQueryTextWithReference()
│   │   ├── Async cache lookup (with fallback)
│   │   ├── Async search operations
│   │   ├── Async LLM prediction
│   │   └── Async cache update
│   └── CompletableFuture.allOf() coordination
└── Final result aggregation
```

### Performance Impact:
- Before: O(n) sequential processing time
- After: O(1) parallel processing time (limited by slowest operation)

### Throughput Check
| NodeNum  | QuerySet Size | SearchConfiguration Size | Judgment Status|
| ------------- | ------------- | ------------- | ------------- |
| 1  | 10  | 5 | COMPLETED |
| 1  | 25  | 5 | COMPLETED |
| 1  | 150  | 5 | COMPLETED |
| 1  | 250  | 5 | COMPLETED |
| 1  | 250  | 10 | COMPLETED |
| 1  | 250  | 15 | COMPLETED |
| 3  | 250  | 5 | COMPLETED |
| 3  | 250  | 10 | COMPLETED |
| 3 | 250  | 15 | COMPLETED |





### Issues Resolved
- #161 
- #159 

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.
For more information on following Developer Certificate of Origin and signing off your commits, please check [here](https://github.com/opensearch-project/OpenSearch/blob/main/CONTRIBUTING.md#developer-certificate-of-origin).